Refactored getActivePage and removed timeout
We should really be always getting the tab
after going through all the pages, and shouldn't have to repeat
