+++ test/utils.js
@@ -27,3 +27,15 @@ export async function initializeExtension() {
     await sleep(100);
     return [browser, service_worker];
 }
+
+export async function getActivePage(browser) {
+    var pages = await browser.pages();
+    var arr = [];
+    for (const p of pages) {
+        if(await p.evaluate(() => { return document.visibilityState == 'visible' })) {
+            arr.push(p);
+        }
+    }
+    if(arr.length == 1) return arr[0];
+    throw "Unable to get active page";
+}
